requirements.txt
streamlit
openai
pytube==15.0.0
openai-whisper
ffmpeg
Pillow
google-generativeai